refactor: extract app api event helpers
